prevent multiple likes

+++ laravel/app/Http/Controllers/IdeaLikeController.php
@@ -10,16 +10,22 @@ class IdeaLikeController extends Controller
     public function like(Idea $id) {
         $liker = auth()->user();
 
-        $liker->likes()->attach($id);
+        if (!$liker->likes->contains($id)) {
+            $liker->likes()->attach($id);
+            return redirect()->back()->with('success', 'Liked successfully!');
+        }
 
-        return redirect()->route('dashboard')->with('success', 'liked successfully!');
+        return redirect()->back()->with('message', 'You have already liked this idea.');
     }
 
     public function unlike(Idea $id) {
         $liker = auth()->user();
 
-        $liker->likes()->detach($id);
+        if ($liker->likes->contains($id)) {
+            $liker->likes()->detach($id);
+            return redirect()->back()->with('success', 'Unliked successfully!');
+        }
 
-        return redirect()->route('dashboard')->with('success', 'unliked successfully!');
+        return redirect()->back()->with('message', 'You have not liked this idea.');
     }
 }
